+++ .github/workflows/id.yml
@@ -8,12 +8,42 @@ on:
     branches: [ main ]
     paths:
       - 'id/**'
+  merge_group:
+    types: [checks_requested]
   workflow_dispatch:
 
 env:
   DATABASE_URL: postgres://postgres:postgres@localhost:5432/noo_id_test
 
 jobs:
+  lint:
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: id
+    steps:
+    - uses: actions/checkout@v4
+    - uses: oven-sh/setup-bun@v2
+      with:
+        bun-version: 1.2.3
+    - name: Install dependencies
+      run: bun install
+    - name: Run all checks
+      run: |
+        set +e
+        FAIL=0
+
+        echo "Running Typecheck..."
+        bun tsc || FAIL=1
+
+        echo "Running Next.js lint..."
+        bun next lint || FAIL=1
+
+        echo "Running Prettier check..."
+        bun prettier --check . || FAIL=1
+
+        exit $FAIL
+
   unit:
     runs-on: ubuntu-latest
     defaults:
@@ -97,7 +127,7 @@ jobs:
 
   build:
     runs-on: ubuntu-24.04-arm
-    needs: [unit, test]
+    needs: [lint, unit, test]
     if: github.ref == 'refs/heads/main'
     permissions:
       packages: write
