--- id/src/app/security/actions.ts
@@ -1,10 +1,0 @@
-"use server";
-
-import Sessions from "@/db/sessions";
-import { redirect } from "next/navigation";
-
-export async function terminateSession(uid: string, sid: string) {
-  await Sessions.destroy(sid);
-
-  return redirect(`/security/sessions?uid=${uid}`);
-}
