+++ id/src/screens/security/sessions/ipUtils.ts
@@ -1,0 +1,3 @@
+export function cleanIp(ip: string) {
+  return ip.replace(/^::ffff:(\d+\.\d+\.\d+\.\d+)$/, "$1");
+}
